Add sudo class `include` and `include_dir` parameters.
Replacing the `include_dirs` parameter with these parameters.  This is
due to the sudo docs saying that the `#includedir` option is used to
determine where the systems package manager will install policy in the
install process, not what the module was using it for. Therefore, an
option to singly define this is added as well as a parameter to pass
multiple files to include in the policy.